add: Created a style for users to click back through their answers in tools
Created a new style that lists the questions and answers to show a users journey through a tool. This allows them to click back to a specific point in a tool and change the answer. An example can be seen here https://www.citizensadvice.org.uk/consumer/energy/energy-supply/problems-with-your-energy-bill/find-out-if-youre-responsible-for-paying-an-energy-bill/