fix: Navigation item focus visibility
Get advice and get involved were losing their focus borders completely on mobile, and partially on desktop. This was fixed by making the spans inline rather than inline-block and removing overflow to hide the outline clipping. It was solved on mobile by adding outline to the span, which is mobile only as it puts a yellow line between the words.